fix: add type definitions, extract hook, and implement security hardening
