+++ package.json
@@ -28,9 +28,11 @@
     "@google/genai": "^1.8.0",
     "@mistralai/mistralai": "^1.7.3",
     "@monaco-editor/react": "^4.6.0",
+    "@types/dompurify": "^3.0.5",
     "@types/styled-components": "^5.1.34",
     "antd": "^5.7.2",
     "core-js": "^3.37.1",
+    "dompurify": "^3.3.1",
     "highlight.js": "^11.10.0",
     "html2pdf.js": "^0.14.0",
     "immer": "^10.1.1",
