+++ src/AgreementHtml.tsx
@@ -3,6 +3,7 @@ import { LoadingOutlined } from "@ant-design/icons";
 import { Spin } from "antd";
 import useAppStore from "./store/store";
 import FullScreenModal from "./components/FullScreenModal";
+import DOMPurify from 'dompurify';
 
 function AgreementHtml({ loading, isModal }: { loading: boolean; isModal?: boolean }) {
   const agreementHtml = useAppStore((state) => state.agreementHtml);
@@ -45,7 +46,7 @@ function AgreementHtml({ loading, isModal }: { loading: boolean; isModal?: boole
       ) : (
         <div
           className="agreement"
-          dangerouslySetInnerHTML={{ __html: agreementHtml }}
+          dangerouslySetInnerHTML={{ __html: DOMPurify.sanitize(agreementHtml) }}
           style={{ flex: 1, color: textColor, backgroundColor: backgroundColor }}
         />
       )}
